Split file up.
