+++ model/src/applications/interaction_option_type.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize)]
+#[serde(untagged)]
+pub enum OptionType {
+    Int(i64),
+    Str(String),
+}
