+++ model/src/applications/mod.rs
@@ -1,12 +1,23 @@
 mod choice;
 mod command;
+mod command_callback_data;
 mod interaction;
+mod interaction_data;
+mod interaction_data_option;
+mod interaction_option_type;
+mod interaction_response;
+mod interaction_response_type;
+mod interaction_type;
 mod option;
 
 pub use choice::*;
 pub use command::*;
-pub use interaction::{
-    CommandCallbackData, Interaction, InteractionDataOption, InteractionResponse,
-    InteractionResponseType, InteractionType,
-};
+pub use command_callback_data::CommandCallbackData;
+pub use interaction::Interaction;
+pub use interaction_data::InteractionData;
+pub use interaction_data_option::InteractionDataOption;
+pub use interaction_option_type::OptionType;
+pub use interaction_response::InteractionResponse;
+pub use interaction_response_type::InteractionResponseType;
+pub use interaction_type::InteractionType;
 pub use option::*;
